[FEATURE] - make setup_logging independent
