+++ syncmaster/backend/middlewares/__init__.py
@@ -4,7 +4,6 @@
 from fastapi import FastAPI
 
 from syncmaster.backend.middlewares.cors import apply_cors_middleware
-from syncmaster.backend.middlewares.logging import setup_logging
 from syncmaster.backend.middlewares.monitoring.metrics import (
     apply_monitoring_metrics_middleware,
 )
@@ -13,6 +12,7 @@ from syncmaster.backend.middlewares.request_id import apply_request_id_middlewar
 from syncmaster.backend.middlewares.session import apply_session_middleware
 from syncmaster.backend.middlewares.static_files import apply_static_files
 from syncmaster.backend.settings import ServerAppSettings as Settings
+from syncmaster.settings.log import setup_logging
 
 
 def apply_middlewares(
